% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_hmd.R
\name{anova_hmd}
\alias{anova_hmd}
\title{Perform anova}
\usage{
anova_hmd(fit)
}
\arguments{
\item{fit}{fit object from GLM model}
}
\value{
Returns fit object of lm function
}
\description{
Perform anova
}
\details{
This functions allows the user to perform anova on the fitted model.
}
