% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_insurance_data.R
\docType{data}
\name{car_insurance_data}
\alias{car_insurance_data}
\title{Car Insurance Data}
\format{
A data frame with 10000 rows and 19 columns:
\describe{
  \item{ID}{Unique ID of the beneficiary.}
  \item{AGE}{Age of the beneficiary, categorize: 26-39, 40-64, other.}
  \item{GENDER}{Gender of the beneficiary, categorized as: male, female.}
  \item{RACE}{Rage of beneficiary, categorized as: major, minor.}
  \item{DRIVING_EXPERIENCE}{Duration of the beneficiary's experience in driving, categorized as:
  0-9y, 10-19y, other.}
  \item{EDUCATION}{Education level of the beneficiary, categorized as: high school, university.}
  \item{INCOME}{Income cass of the beneficiary, categorized as: upper class, middle class, other.}
  \item{CREDIT_SCORE}{Credit score of beneficiary.}
  \item{VEHICLE_OWNERSHIP}{Whether the beneficiary owns the vehicle or not.}
  \item{VEHICLE_YEAR}{The year vehicle was built, categorized as: before 2015, after 2015.}
  \item{MARRIED}{Whether the beneficiary is married or not.}
  \item{CHILDREN}{Whether the beneficiary has children or not.}
  \item{POSTAL_CODE}{Postal code of the beneficiary.}
  \item{VEHICLE_TYPE}{The type of the vehicle of the beneficiary, categorized as: sedan, sports car.}
  \item{ANNUAL_MILEAGE}{Annual mileage of the vehicle of the beneficiary.}
  \item{SPEEDING_VIOLATIONS}{Number of speeing violations recorded for beneficiary}
  \item{DUIS}{Number of times beneficiary did driving under the influence (DUI).}
  \item{PAST_ACCIDENTS}{Number of beneficiary's past driving accidents.}
  \item{OUTCOME}{Indicates 1 if a customer has claimed his/her loan else 0.}
}
}
\source{
<https://www.kaggle.com/datasets/sagnik1511/car-insurance-data>
}
\usage{
data(car_insurance_data)
}
\description{
This dataset includes annual car insurance data shared by a company. Most of
the data is real and part of it is artificially generated. The target column is
`outcome`. This binary column shows if the insurance request was rejected or accepted.
}
\keyword{datasets}
