% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricals_hmd.R, R/categoricals_print_hmd.R,
%   R/prprocess_hmd.R
\name{categoricals_hmd}
\alias{categoricals_hmd}
\title{Treat categorical columns}
\usage{
categoricals_hmd(data = regPackage1::car_insurance_data)

categoricals_hmd(data = regPackage1::car_insurance_data)

categoricals_hmd(data = regPackage1::car_insurance_data)
}
\arguments{
\item{data}{An arbitrary dataframe}
}
\value{
Returns dataframe with its categorical columns factorised

Returns cleaned dataframe ready to be fed into models
}
\description{
Treat categorical columns

Print nature of categorical columns in a given dataframe

Prepare dataframe to bed fed into models
}
\details{
This functions treats the categorical columns by first detecting them and the
convert them to factors. To detect categorical columns, the function selects a column
as categorical if either its class is "character", or if number of unique values
the column has is less than 5.

This functions first detects which columns are categorical and then prints whether
the column contains numbers or characters. To detect categorical columns, the function selects a column
as categorical if either its class is "character", or if number of unique values
the column has is less than 5.

This functions allows to clean and prepare dataframe by cleaning column names, automatically detecting the
categorical columns and then factorizing them, and replace "na" and "Inf" values with NA
}
