% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricals_print_hmd.R
\name{categoricals_print_hmd}
\alias{categoricals_print_hmd}
\title{Print nature of categorical columns in a given dataframe}
\usage{
categoricals_print_hmd(data = regPackage1::car_insurance_data)
}
\arguments{
\item{data}{An arbitrary dataframe}
}
\description{
Print nature of categorical columns in a given dataframe
}
\details{
This functions first detects which columns are categorical and then prints whether
the column contains numbers or characters. To detect categorical columns, the function selects a column
as categorical if either its class is "character", or if number of unique values
the column has is less than 5.
}
