% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hmd.R
\name{eval_hmd}
\alias{eval_hmd}
\title{Evaluation Metrics Computed}
\usage{
eval_hmd(actual, predicted, predict_proba)
}
\arguments{
\item{actual}{The actual data (usually obtained from test data)}

\item{predicted}{Predictions from an arbitrary model}

\item{predict_proba}{Prediction probablities from a classification model (e.g. logistic regression)
except the target}
}
\value{
Returns a hash of all evaluation metrics
}
\description{
Evaluation Metrics Computed
}
\details{
This functions computes various evaluation metrics of classification for a given
actual data and a prediction data. The function returns a hash containing the value of metrics
}
