% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_hmd.R
\name{glm_hmd}
\alias{glm_hmd}
\title{GLM Model}
\usage{
glm_hmd(
  data = insurance_data,
  features_names = names({
     {
         data
     }
 })[names({
     {
         data
 
       }
 }) != {
     {
         target
     }
 }],
  target = "charges",
  family = "gaussian"
)
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{features_names}{Name of features. If not specified, takes all columns
except the target}

\item{target}{The target variable aimed for prediction}

\item{family}{specify family of distribution.}
}
\value{
Returns fit object of glm function
}
\description{
GLM Model
}
\details{
This functions allows the user to perform generalized linear model on a given
dataframe by specifying feature names (response variables),
target variable, family of distribution, and the dataset (in my case, Medical Cost Personal Datasets)
}
