% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_plot_hmd.R
\name{glm_plot_hmd}
\alias{glm_plot_hmd}
\title{Plot avPlots of GLM model}
\usage{
glm_plot_hmd(fit)
}
\arguments{
\item{fit}{fit object from GLM}
}
\value{
Returns avPlot object from the car package
}
\description{
Plot avPlots of GLM model
}
\details{
Plot an added-variable plot of the dependent variable and one independent
variable from a multiple regression.
}
