% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_cv_aic_hmd.R
\name{glmnet_cv_aic_hmd}
\alias{glmnet_cv_aic_hmd}
\title{Predict data with results from GLMNET model}
\usage{
glmnet_cv_aic_hmd(fit, lchoice = "min")
}
\arguments{
\item{fit}{fit object from a cv.glmnet model}

\item{lchoice}{how to choose lambda model. Choices are "min", "1se".
for "binomial" or "multinomial" models; for "gaussian" models it gives the fitted values.
Type "response" gi}
}
\value{
Returns fit object of glmnet function
}
\description{
Predict data with results from GLMNET model
}
\details{
This functions allows the user to perform cross-validated eslastic net (which
is a generalized
linear model) on a given
dataframe by specifying feature names (response variables),
target variable, family of distribution, and the dataset (in my case, Medical
Cost Personal Datasets)
}
