% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_cv_hmd.R
\name{glmnet_cv_hmd}
\alias{glmnet_cv_hmd}
\title{CV GLMNET model}
\usage{
glmnet_cv_hmd(
  data = car_insurance_data,
  features_names = names({
     {
         data
     }
 })[names({
     {
         data
 
       }
 }) != {
     {
         target
     }
 }],
  target = "outcome",
  family = "poisson"
)
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{features_names}{Name of features. If not specified, takes all columns
except the target}

\item{target}{The target variable aimed for prediction}

\item{family}{specify family of distribution.}
}
\value{
Returns fit object of glmnet function
}
\description{
CV GLMNET model
}
\details{
This functions allows the user to perform cross-validated eslastic net (which is a generalized
linear model) on a given
dataframe by specifying feature names (response variables),
target variable, family of distribution, and the dataset (in my case, Medical Cost Personal Datasets)
}
