% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_cv_predict_hmd.R
\name{glmnet_cv_predict_hmd}
\alias{glmnet_cv_predict_hmd}
\title{Predict data with results from CV.GLMNET model}
\usage{
glmnet_cv_predict_hmd(
  fit,
  data = regPackage1::insurance_test,
  target = "outcome",
  lchoice = "min",
  type = "binomial"
)
}
\arguments{
\item{fit}{fit object from a cv.glmnet model}

\item{data}{An arbitrary dataframe}

\item{target}{Target column in the data}

\item{lchoice}{how to choose lambda model. Choices are "min", "1se".}

\item{type}{Type of prediction required. Type "link" gives the linear predictors
for "binomial" or "multinomial" models; for "gaussian" models it gives the fitted values.
Type "response" gi}
}
\value{
Returns prediction of the cv.glmnet function
}
\description{
Predict data with results from CV.GLMNET model
}
\details{
This functions allows the user to predict a given dataset using fit object
obtained from the cv.glmnet function
}
