% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_predict_hmd.R
\name{glmnet_predict_hmd}
\alias{glmnet_predict_hmd}
\title{Predict data with results from GLMNET model}
\usage{
glmnet_predict_hmd(
  fit,
  data = regPackage1::insurance_test,
  target = "outcome",
  type = "binomial"
)
}
\arguments{
\item{fit}{fit object from a cv.glmnet model}

\item{data}{An arbitrary dataframe}

\item{target}{Target column in the data}

\item{type}{Type of prediction required. Type "link" gives the linear predictors
for "binomial" or "multinomial" models; for "gaussian" models it gives the fitted values.
Type "response" gi}
}
\value{
Returns fit object of glmnet function
}
\description{
Predict data with results from GLMNET model
}
\details{
This functions allows the user to perform cross-validated eslastic net (which
is a generalized
linear model) on a given
dataframe by specifying feature names (response variables),
target variable, family of distribution, and the dataset (in my case, Medical
Cost Personal Datasets)
}
