% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_median_hmd.R
\name{impute_median_hmd}
\alias{impute_median_hmd}
\title{Imputes missing values with median}
\usage{
impute_median_hmd(data = regPackage1::car_insurance_data, cols_list = list())
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{cols_list}{List of numerical columns intended to be imputed}
}
\value{
Returns dataframe with its categorical columns factorised
}
\description{
Imputes missing values with median
}
\details{
This function receives a dataframe and a list of columns to be imputed with median.
Then it replaces NA values of each column from the given list with median of that column.
}
