% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_data.R
\docType{data}
\name{insurance_data}
\alias{insurance_data}
\title{Medical Cost Personal Datasets}
\format{
A data frame with 1,338 rows and 7 columns:
\describe{
  \item{age}{Age of primary beneficiary}
  \item{sex}{Insurance contractor gender, female, male}
  \item{bmi}{Body mass index, providing an understanding of body, weights that
  are relatively high or low relative to height, objective index of body weight $(kg / m ^ 2)$ using the ratio of height to weight, ideally 18.5 to 24.9}
  \item{children}{Number of children covered by health insurance / Number of dependents}
  \item{smoker}{Individual is smoker or not}
  \item{region}{The beneficiary's residential area in the US, northeast, southeast, southwest, northwest.}
  \item{charges}{Individual medical costs billed by health insurance}
}
}
\source{
<https://www.kaggle.com/datasets/mirichoi0218/insurance>
}
\usage{
data(insurance_data)
}
\description{
Individual medical costs billed by health insurance along with properties of individuals
}
\keyword{datasets}
