% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_glmnet_cv_hmd.R
\name{predict_glmnet_cv_hmd}
\alias{predict_glmnet_cv_hmd}
\title{Predict data with results from CV GLMNET model}
\usage{
predict_glmnet_cv_hmd(fit, data, lchoice = "min", type = "link")
}
\arguments{
\item{fit}{Fit object from a cv.glmnet model}

\item{data}{An arbitrary dataframe}

\item{lchoice}{How to choose lambda model. Choices are "min", "1se"}

\item{type}{Type of prediction required. Type "link" gives the linear
predictors for "binomial" or "multinomial" models;
for "gaussian" models it gives the fitted values. Type "response" gi.}
}
\value{
Returns fit object of glmnet function
}
\description{
Predict data with results from CV GLMNET model
}
\details{
This functions allows the user to perform cross-validated eslastic net (which
is a generalized
linear model) on a given
dataframe by specifying feature names (response variables),
target variable, family of distribution, and the dataset (in my case, Medical
Cost Personal Datasets)
}
