% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hmd.R
\name{prepare_hmd}
\alias{prepare_hmd}
\title{Fitting a generalized linear model by specifying feature names (response variables),
target variable, family of distribution}
\usage{
prepare_hmd(data = insurance_data, cat_cols = list())
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{cat_cols}{List of cateogrical columns need to be factored and result in levels}
}
\value{
Returns fit object of lm function
}
\description{
Fitting a generalized linear model by specifying feature names (response variables),
target variable, family of distribution
}
\details{
This functions allows the user to perform generalized linear model on a given
dataframe, in my case, Medical Cost Personal Datasets
}
