% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prprocess_hmd.R
\name{preprocess}
\alias{preprocess}
\title{Prepare dataframe to bed fed into models}
\usage{
preprocess(data = regPackage1::car_insurance_data)
}
\arguments{
\item{data}{An arbitrary dataframe}
}
\value{
Returns cleaned dataframe ready to be fed into models
}
\description{
Prepare dataframe to bed fed into models
}
\details{
This functions allows to clean and prepare dataframe by cleaning column names, automatically detecting the
categorical columns and then factorizing them, and replace "na" and "Inf" values with NA
}
