% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_fit_hmd.R
\name{rf_fit_hmd}
\alias{rf_fit_hmd}
\title{Random Forest Model Fitting}
\usage{
rf_fit_hmd(
  data = regPackage1::insurance_train,
  target = "outcome",
  family = "binomial",
  ntree = 300,
  mtry = 10,
  proximity = TRUE,
  importance = TRUE
)
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{target}{The target variable aimed for prediction}

\item{family}{Specify family of distribution.}

\item{ntree}{Number of trees to grow. This should not be set to too small a '
umber, to ensure that every input row gets predicted at least a few times.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.
Note that the default values are different for classification (sqrt(p) where p is number of variables in x) and regression (p/3)}

\item{proximity}{Should proximity measure among the rows be calculated?}

\item{importance}{Should importance of predictors be assessed?}
}
\value{
Returns fit object of randomForest function
}
\description{
Random Forest Model Fitting
}
\details{
This functions allows the user to fit a random forest on a given dataframe
and to choose the hyperparameters for it.
}
