% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_predict.R
\name{rf_predict_hmd}
\alias{rf_predict_hmd}
\title{Predict data with results from random forest model}
\usage{
rf_predict_hmd(fit, data = regPackage1::insurance_test, target = "outcome")
}
\arguments{
\item{fit}{fit object from a randomfForest model}

\item{data}{An arbitrary dataframe}

\item{target}{Target column in the data}
}
\value{
A hash containing predictions, and prediction probabilities
}
\description{
Predict data with results from random forest model
}
\details{
This functions allows the user to predict a given dataset using fit object
obtained from the randomForest function
}
