% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_splitter_hmd.R
\name{splitter_norm_hmd}
\alias{splitter_norm_hmd}
\title{Splitting and Normalizind Data}
\usage{
splitter_norm_hmd(data = regPackage1::car_insurance_data, proportion = 0.7)
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{proportion}{Proportion of train to test. Default is 0.7, which results}
}
\value{
Returns a dictionary containing train and test data
}
\description{
Splitting and Normalizind Data
}
\details{
This functions allows the user to split data into train and test
}
