% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_hmd.R
\name{anova_hmd}
\alias{anova_hmd}
\title{Fitting a generalized linear model by specifying feature names (response variables),
target variable, family of distribution}
\usage{
anova_hmd(fit)
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{features_names}{Name of features. If not specified, takes all columns
except the target}

\item{target}{The target variable aimed for prediction}

\item{normalize}{If set to TRUE, then the features are normalized either by
minmax or std, which is determined by the parameter "normalized_method"}

\item{normalize_method}{if set to "std", then standard scaler is used for
normalization of features, and if set to "minmax", then minmax scaler is used. Default is "std".}

\item{family}{specify family of distribution.}
}
\value{
Returns fit object of lm function
}
\description{
Fitting a generalized linear model by specifying feature names (response variables),
target variable, family of distribution
}
\details{
This functions allows the user to perform generalized linear model on a given
dataframe, in my case, Medical Cost Personal Datasets
}
