% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricals_hmd.R
\name{categoricals_hmd}
\alias{categoricals_hmd}
\title{Treat categorical columns}
\usage{
categoricals_hmd(data = regPackage1::car_insurance_data)
}
\arguments{
\item{data}{An arbitrary dataframe}
}
\value{
Returns dataframe with its categorical columns factorised
}
\description{
Treat categorical columns
}
\details{
This functions treats the categorical columns by first detecting them and then
convert those including characters to factors. To detect categorical columns, the function selects a column
as categorical if either its class is "character", or if number of unique values
the column has is less than 5.
}
