% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_hmd.R
\name{glm_hmd}
\alias{glm_hmd}
\title{Fitting a generalized linear model by specifying feature names (response variables),
target variable, family of distribution}
\usage{
glm_hmd(
  data = insurance_data,
  features_names = names({
     {
         data
     }
 })[names({
     {
         data
 
       }
 }) != {
     {
         target
     }
 }],
  target = "charges",
  normalize = FALSE,
  normalize_method = "std",
  family = "gaussian"
)
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{features_names}{Name of features. If not specified, takes all columns
except the target}

\item{target}{The target variable aimed for prediction}

\item{normalize}{If set to TRUE, then the features are normalized either by
minmax or std, which is determined by the parameter "normalized_method"}

\item{normalize_method}{if set to "std", then standard scaler is used for
normalization of features, and if set to "minmax", then minmax scaler is used. Default is "std".}

\item{family}{specify family of distribution.}
}
\value{
Returns fit object of lm function
}
\description{
Fitting a generalized linear model by specifying feature names (response variables),
target variable, family of distribution
}
\details{
This functions allows the user to perform generalized linear model on a given
dataframe, in my case, Medical Cost Personal Datasets
}
