% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_median_hmd.R
\name{impute_median_hmd}
\alias{impute_median_hmd}
\title{Imputes missing values with median}
\usage{
impute_median_hmd(data = regPackage1::car_insurance_data)
}
\arguments{
\item{data}{An arbitrary dataframe}
}
\value{
Returns dataframe with each missing value is replaced by the median of the
column in which the missing value exists.
}
\description{
Imputes missing values with median
}
\details{
This function receives a dataframe, finds columns containing NA values, and then
replaces NA values of each found column with median of that column.
}
