% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hmd.R
\name{prepare_hmd}
\alias{prepare_hmd}
\title{Prepare dataframe to bed fed into models
target variable, family of distribution}
\usage{
prepare_hmd(data = regPackage1::car_insurance_data)
}
\arguments{
\item{data}{An arbitrary dataframe}
}
\value{
Returns cleaned dataframe ready to be fed into models
}
\description{
Prepare dataframe to bed fed into models
target variable, family of distribution
}
\details{
This functions allows to clean and prepare dataframe by cleaning column names, automatically detecting the
categorical columns and then factorizing them, and replace "na" and "Inf" values with NA
}
