% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hmd.R
\name{prepare_hmd}
\alias{prepare_hmd}
\title{Prepare dataframe to bed fed into models
target variable, family of distribution}
\usage{
prepare_hmd(data = car_insurance_data, cat_cols = list())
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{cat_cols}{Set of categorical columns to be factorized}
}
\value{
Returns cleaned dataframe ready to be fed into models
}
\description{
Prepare dataframe to bed fed into models
target variable, family of distribution
}
\details{
This functions allows to clean and prepare dataframe by factorizing the
specified cateogrical columns, cleaning column names, replace "na" and "Inf"
values with NA
}
