% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_hmd.R
\name{preprocess_hmd}
\alias{preprocess_hmd}
\title{Prepare dataframe to be fed into models}
\usage{
preprocess_hmd(data = regPackage1::car_insurance_data)
}
\arguments{
\item{data}{An arbitrary dataframe}
}
\value{
Returns cleaned dataframe ready to be fed into models
}
\description{
Prepare dataframe to be fed into models
}
\details{
This function is a wrapper around all functions defined in the preprocessing stage.
It allows the user to use the raw original data and using this function receive a data
ready to be fed into models. This function applies the following functions on the given dataframe:
janitor::clean_names, regPackage1::categoricals_hmd, regPackage1::impute_median_hmd
}
