% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter_norm_hmd.R
\name{splitter_norm_hmd}
\alias{splitter_norm_hmd}
\title{Splitting and Normalizind Data}
\usage{
splitter_norm_hmd(
  data = insurance_data,
  features_names = names({
     {
         data
     }
 })[names({
     {
         data
 
       }
 }) != {
     {
         target
     }
 }],
  target = "charges",
  normalize = FALSE,
  normalize_method = "std"
)
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{features_names}{Name of features. If not specified, takes all columns
except the target}

\item{target}{The target variable aimed for prediction}

\item{normalize}{If set to TRUE, then the features are normalized either by
minmax or std, which is determined by the parameter "normalized_method"}

\item{normalize_method}{if set to "std", then standard scaler is used for
normalization of features, and if set to "minmax", then minmax scaler is used. Default is "std".}
}
\value{
Returns a list of containing train and test data (both normalized if user requests)
}
\description{
Splitting and Normalizind Data
}
\details{
This functions allows the user to split data into train and test, and then
if the user sets normalize to TRUE, it normalizes data based on the moethod
user prefers
}
