% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter_norm_hmd.R
\name{splitter_norm_hmd}
\alias{splitter_norm_hmd}
\title{Splitting and Normalizind Data}
\usage{
splitter_norm_hmd(
  data = regPackage1::car_insurance_data,
  replace = TRUE,
  proportion = 0.7,
  normalize = FALSE,
  normalize_method = "minmax"
)
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{replace}{If set to TRUE, the split will be replaced samples}

\item{proportion}{Proportion of train to test. Default is 0.7, which results
in 935 rows for train and 403 rows for test in case of using insurance_data.}

\item{normalize}{If set to TRUE, then the features are normalized either by
minmax or std, which is determined by the parameter "normalize_method"}

\item{normalize_method}{if set to "std", then standard scaler is used, and if
set to "minmax", minmax is applied.
normalization of features, and if set to "minmax", then minmax scaler is used. Default is "std".}
}
\value{
Returns a list of containing train and test data (both normalized if user requests)
}
\description{
Splitting and Normalizind Data
}
\details{
This functions allows the user to split data into train and test, and then
if the user sets normalize to TRUE, it normalizes data based on the moethod
user prefers
}
